Add files via upload
Adding the python module that calls the Google Cloud Vision API and takes in the image. The image uploaded is simply a test image, a placeholder.
The module's code will be changed in a bit, I will incorporate the pandas library to output the data in a nicer more useful way.
See comments for more information.